Adds inception.utils to inception package.
Previously inception.utils was overlooked for packaging.

Note: the use of setuptools' "find_packages()" might appear to be
a poor choice since this setup.py is written to fall back to using
distutils which doesn't have a find_packages but in fact it cannot
usefully fall back to distutils anyway (see bug #1210181) so it is
not an issue.

Change-Id: I31b2b5f3952df0b8ecd0bc59f9978cf0bb9f03e5
Fixes: bug #1210180