Provide simple RESTful API + WSGI Server
blueprint horizon-webui

(Auto-linking to blueprint either doesn't work for this project
or requires additional info not prescribed. Here is full link:
https://blueprints.launchpad.net/inception/+spec/horizon-webui)

This commit implements the minimal RESTful API and WSGI application
for exposing Inception Cloud's Orchestrator behavior over the network.
It is a necessary component for providing a web user interface to
Orchestrator via OpenStack's horizon project as described in the
Blueprint identified above.

Change-Id: I8430b008c6cbabc91258da31d01a05a5521462d0