Illustrations are now generated into PDF files to support older versions of tex
Versions of LaTex prior to 2010 do not supprot the inclusion of encapsulated
postscript files.  Changes have been made to generate PDF files for each of
the illustrations in the document, clean up these temporary PDF files when
`make clean` is executed, and to embed them rather than .eps files when the
document is compiled.

Change-Id: I2a802a82cb0ead4eb379f81f0a1ed441ce2967f1