Supply horizon-compatible Web UI
blueprint horizon-webui

(Auto-linking to blueprint either doesn't work for this project
or requires additional info not prescribed. Here is full link:
https://blueprints.launchpad.net/inception/+spec/horizon-webui

This change is the initial commit of the third and final installment
of the horizon-compatible web UI for Inception Cloud (the others are
[1] and [2]).  This commit draws heavily from the horizon tutorial[3]
and from the existing implementation for instance management[4] (the
latter because inception cloud management is by analogy a more complicated
version of it).

Because this effort seeks to operate alongside OpenStack in the
architectural and coding style of OpenStack but require minimal internal
changes to OpenStack it relies on a separate API server from nova
compute.  In addition, this meant that the entire horizon stack could not
be utilized directly and a layer had to be chosen at which to deviate.
Accordingly the code is more verbose (because of horizon generality) than
it might otherwise be.

Installation of this application proceeds by
1. installing the inception package
2. starting the inception API server per INSTALL.md
3. modifying a single line in the horizon config per INSTALL.md

[1] https://review.openstack.org/#/c/47008/
[2] https://review.openstack.org/#/c/58835/
[3] http://docs.openstack.org/developer/horizon/topics/tutorial.html
[4] https://github.com/openstack/horizon/ then
    openstack_dashboard/dashboards/project/instances

Partially implements: blueprint horizon-webui

Change-Id: Id14b30bbb5eafeac928d8e94aeb0553ae52fb661